backend/requirements.txt
Django
djangorestframework
gunicorn
psycopg2-binary
whitenoise
django-cors-headers
djangorestframework-simplejwt
dj-database-url==3.0.1
django-q2
